Remove dependency on ghc API package
This dependency has gone unnoticed for a long time (and was probably mistaken for a way to specify the ghc version.) Without it we will probably get fewer dependency problems as the prototype depends on fewer packages that are linked directly to the ghc version used. Note that it is still dangerous to not have package upper bounds, as a clean install of Ampersand may fail if one of the dependencies gets an update (you won't notice this in your current installation).

git-svn-id: http://svnint.ou.nl:8080/svn/ADL/trunk@3222 80d59baa-9aef-c140-ba76-e18f3b6eea4d