Fix all dependencies on minor version
This way Ampersand cannot become unbuildable when a dependency is updated. Without an upper bound, a clean install may choose a newer version of a dependency, which may install a newer version of a dependency shared with one of Ampersand's other dependencies, which will then create a conflict. Using the topDown solver helps for Ampersand, but will still create problems with the prototype.

git-svn-id: svn+ssh://svn.code.sf.net/p/ampersand/code/trunk@1439 8b35ee93-31c1-4acc-b283-6c2059e6c15c