Fix all dependencies on minor version
This way the prototype cannot become unbuildable when a dependency is updated. Without an upper bound, a clean install may choose a newer version of a dependency, which may install a newer version of a dependency shared with one of Ampersand's other dependencies, which will then create a conflict. Using the topDown solver helps for Ampersand, but will still create problems with the prototype.

git-svn-id: http://svnint.ou.nl:8080/svn/ADL/trunk@3233 80d59baa-9aef-c140-ba76-e18f3b6eea4d