Move getEditableRelation to ampersand
We could also choose to store the result of this function in the interface object data structure (ie. ObjectDef), but since this is an A-datatype rather than an Fspec datatype, this would require quite some refactoring to prevent import cycles. It is also debatable whether this kind of information should be part of the A-structure (storing it only in the Fspec is not doable as we would need to copy the tree structure or do some nasty kind of lookup).