Remove distinction between binary and textual static files
On Windows, linefeed-only files are handled well enough (and won't even appear if git autocrlf=true is used,) so there is no need for special treatment of textual files. All static files in generated prototypes are now exact copies of their counterparts in 'ampersand/static/'. The simplification will make it much easier to use the static mechanism for larger directories.