Read previously generated statics module and only update on changes
The added complexity is necessary to handle the large amount of static files for the new prototype. Compiling StaticFiles_Generated takes 10 seconds on a fast machine, which is not a feasible solution unless we only need to do it on changes to those static files. Once cabal handles recursive patterns in data-files, we can get rid of StaticFiles_Generated altogether and use data-files for all static files.