Add support for user-defined templates
Since specific templates are essentially part of the context definition (just like the old Custom.css), you can create a 'templates' subdirectory in the directory of the .adl context, and any templates in it will override the default templates. The directory structure follows the structure of static/newFrontend/templates, with views and controllers subdirecties.