Undo bump mtl-2.1.* to 2.2.* + Sort dependencies
Despite the remark in [60c9c13], mtl-2.2.* is not actually required for simle-sql-parser-0.4.1. Using mtl-2.1.* saves us a problem on Linux, where ghc has regex-posix and regex-compat installed by default. Package mtl-2.2 causes a reinstall of regex-base (because in a sandbox only a single version of each package is allowed), which yields an error that regex-posix and regex-compat will be broken unless --force-reinstalls is provided.