promote a comment of Rieks to a pattern purpose
Yes, Rieks, this type of comment is useful. Martijn has also made similar remarks in the past. I will leave choosing more appropriate names to Han too, because we are (manually!) keeping the Haskell code and AST in sync. So that has to be done with some care. Occasionally I will redo a name myself, if I have the opportunity.