provide an example for ticket #202
Try running Atoms.ifc (which includes Atoms.adl and Atoms.xlsx).
This is the preferred implementation according to 2 relation-algebra specialists at the Braga conference, who regard the function repr as undesired in this program. Indeed. the program works as desired and has become much simpler. You can try it by adding a pair. The only thing that keeps it from running completely is Error! 501 Not Implemented: TgtConcept of interface: 'atoms' is scalar and can not have subinterfaces.