Make haddock work with quick hack
Haddock currently fails to run, failing on "haddock library 1.4.5" which
is included by the current version of pandock. As far as my googling
tells me, this would be fixed once we can upgrade to "haddock library
1.6.0". I've been told the hold-back is "pandock-crossref".

A better fix would be to only exclude the problematic package from
haddock, but I could not figure out if that is possible.