Create basic framework for pre-processor.
The parser is changed to read flags from an include (Read as an optional
bracketed list). These options are propagated to a ParseCandidate, where
previous options are merged in. (Currently merge just concatenates list.
Will need to deal with turning off options later).

Before we run the actual parser on a file, we run the preprocessor with
the given options. Alternatively for performance, we could have the
lexer do the pre-processing.