First attempt at PreProcessor.
Swapped arguments on PreProcess to make it point-free.
Current parser implementation is very ugly, and might not work.

Also, the Include statement as is takes values that are quoted
as strings, this is not elegant. Moreover, it currently takes the
arguments in 'real code' rather than in comments. This should be changed
to retain backwards compatibility for code that uses the pre-processor.