Change preprocessor to use Parsec.
Technically, the lexer could also be replaced, but the current version
seems fine. Perhaps if we want more flexibility in the commands, or
want to retain comments that are after the guard in the output.

The function turning a block back into a (partially commented)
string could probably also be replaced by a stateful parser. Not at all
necessary though.

Finally, we should return something like Either ParseError String from
preProcess so we can propagate parsing errors rather than crashing.