Refactor entire preprocessor, minor fixes.
List of fixes:
- We can now handle files that aren't terminated by a newline
  (we do this by crudely appending a newline to each file)
- We expanded the syntax for a preprocessor directive.
  Now, we also accept more dashes than just "--" and whitespace between
  the dashes and the '#'.

The most notable refactor is merging the IfBlock and IfNotBlock types
into the GuardedBlock type. This was done to factor out the function
showGuardedBlock, which render both IF and IFNOT blocks. As the logic is
complicated and very similar, it made sense to merge them.

We also reordered the definitions in the Lexer and Parser so the
high-level definitions come first. Finally, we factored out some common
code and added a whole lot of documentation.