Put INCLUDE flags behind an optional "--#"
Reverts the previous commit.

We now allow an include statement with flagsin either of the following 2 forms:
INCLUDE "filename" ["FLAG"]
INCLUDE "filename" --# ["FLAG"]
This is done by the preprocessor detecting the second form and converting
it to the first.

The last commit did not work. There we tried to allow for comments in
the ampersand parser. This actually requires work in the lexer.
Specifically, it requires making the lexer distinguish --# from --.
As this is difficult, this functionality was pulled into the preProcessor.

In the end, we might want to put this, combined with giving the actual flags
in the Lexer.