Fix bug with line-numbers.
There were two issues with line numbers:

* empty Lines before an --#IF or an --#IFNOT were swallowed
* INCLUDE statements without flags (i.e. without a --# on the line) had
an extra newline behind them.

The first issue was because Parsec's `spaces` also accepts newlines as
whitespace. The second issue was due to the complexity of lexing an
include statement. It should not include the newline, but it did in the
case where there are no include flags. This meant an extra newline was
output.