Feature: allow disabling preprocessor flags.
When file A imports file B with flag FOO, then B can import file C with
flag !FOO to ensure flag FOO is not enabled in file C.

Before this change, imports of imports would unconditionally receive all
flags of their parent imports. Note that technically, one can use the
preprocessor to comment out imports. Because reading imports isn't done
by the preprocessor. Hence, one could do:

    --# IF DISABLE_IMPORT
        IMPORT bar --# ["!FOO", "OTHERFLAG"]
    --# ELSE
        IMPORT bar --# ["OTHERFLAG"]
    --# ENDIF

In support of this chance, we refactored aswell. In an import-candidate,
we now use a ```Data.Set``` to store all enabled flags. This is
conceptually better, because we only care about which flags are enabled,
and not their order or multiplicity.

Thus, this commit also touches ```Input/Parsing.hs```.