Fix preProcessor set and unset flags were reversed
Setting flags was treated as unsetting flags, and vice-versa.

That is: ```INCLUDE "foo.adl" --# ["FLAG"]``` was treated like
```INCLUDE "foo.adl" --# ["!FLAG"]```. Notably this meant that almost
all scripts were pre-processed as if no flags were set.

The old test did not catch this. This commit modifies the test to fail
if the pre-processor does not read any flags.