rename expressions to term
In the theory, relations are combined into terms (not in expressions)
In the parser and in the user manual of Ampersand we talk about "terms"
In the A-structure, we talk about "expressions". (that is: the data structure Term and Expression occur both in the compiler)
I have sanitized the use of these words throughout the code and the documentation.